requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
langchain==0.2.7
langchain-community==0.2.7
langchain-core==0.2.11
langchain-openai==0.0.9
chromadb==0.5.3
openai==1.35.13
tiktoken==0.7.0